+++ src/pytest_benchmark/report_backend/elasticsearch_report_backend.py
@@ -1,15 +1,16 @@
+
 from .base_report_backend import BaseReportBackend
 from ..elasticsearch_storage import ElasticsearchStorage
 
 
 class ElasticReportBackend(BaseReportBackend):
     def __init__(self, config):
-        self.elasticsearch_host = config.getoption("benchmark_elasticsearch_host")
+        self.elasticsearch_hosts = config.getoption("benchmark_elasticsearch_hosts")
         self.elasticsearch_index = config.getoption("benchmark_elasticsearch_index")
         self.elasticsearch_doctype = config.getoption("benchmark_elasticsearch_doctype")
         self.project_name = config.getoption("benchmark_project")
         super(ElasticReportBackend, self).__init__(config)
-        self.storage = ElasticsearchStorage(self.elasticsearch_host,
+        self.storage = ElasticsearchStorage(self.elasticsearch_hosts,
                                             self.elasticsearch_index,
                                             self.elasticsearch_doctype,
                                             self.logger,
@@ -41,7 +42,7 @@ class ElasticReportBackend(BaseReportBackend):
                 self.storage.save(bench, doc_id)
             self.logger.info("Saved benchmark data to %s to index %s as doctype %s" %
                              (
-                                 self.elasticsearch_host,
+                                 self.elasticsearch_hosts,
                                  self.elasticsearch_index,
                                  self.elasticsearch_doctype
                              )